Added configuration files supporting the Keshmesh coding conventions.
Added the template file for automatically inserting the license header.
Added the script adapted by Nick Chen for checking the format of commit messages.
Cleaned up the dependencies of the Eclipse projects.
Changed the extension of the license notice.
Moved the files corresponding to the coding conventions to the directory called "coding-conventions".