Fixed a WALA compilation problem in test number 11 of LCK02J.
Removed the @Override annotation from the declaration of B.run because B implements Runnable and doesn't extend a class with method run. Having removed the redundant @Override annotation, WALA detects the instance of LCK02J in the code. But, it only reports p.B.class as what getLock() may point to and not p.C.class.

This commit belongs to issue #35.