Made the context sensitivity of the analysis lighter.
Used K object context sensitivity for non-JDK classes instead of infinite object
context sensitivity.

This change seems to have caused some of the LCK06J and VNA00J tests to fail.
But, it isn't clear how many of the failures are due to this change.

This commit belongs to issue #37.