Added a plug-in that hooks Keshmesh into Findbugs for debugging.
To be able to debug Keshmesh, we need to use the FindBugs extension point
differently. I have added the plug-in edu.illinois.keshmesh.ui.findbugs.debug
just for this purpose. When debugging Keshmesh, we should use this plug-in
instead of edu.illinois.keshmesh.ui.findbugs. That is, we should close plug-in
edu.illinois.keshmesh.ui.findbugs. However, the Keshmesh update site will be
created using edu.illinois.keshmesh.ui.findbugs.

This commit belongs to issue #70.